User forms refact: Both User Update and User Add forms inherit the same base user form dialog class. now we're DRY
