+++ src/app/app.module.ts
@@ -40,6 +40,7 @@ import { UserComponent } from './pages/user/user.component';
 import { AddUserDialogComponent } from './pages/user/dialogs/add/add-user.dialog';
 import { UserService } from './pages/user/services/user.service';
 import { BrowserModule } from '@angular/platform-browser';
+import { UserDetailsFormComponent } from './pages/user/dialogs/base/user-details-form/user-details-form';
 import { DeleteUserDialogComponent } from './pages/user/dialogs/delete/delete-user.dialog';
 import { UpdateUserDialogComponent } from './pages/user/dialogs/update/update-user.dialog';
 import { DeleteStudentDialogComponent } from './pages/student/dialogs/delete/delete-student.dialog';
@@ -90,6 +91,7 @@ import { MSWSnackbar } from './services/msw-snackbar/msw-snackbar.service';
     ClassComponent,
     StudentComponent,
     UserComponent,
+    UserDetailsFormComponent,
     AddUserDialogComponent,
     UpdateUserDialogComponent,
     DeleteUserDialogComponent,
