+++ src/app/pages/user/dialogs/add/add-user.dialog.ts
@@ -1,3 +1,4 @@
+import { BaseUserFormDialogComponent } from './../base/base-user-form-dialog';
 import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
@@ -11,58 +12,8 @@ import { Class } from '../../../../models/class.model';
   templateUrl: './add-user.dialog.html',
   styleUrls: ['./add-user.dialog.scss'],
 })
-export class AddUserDialogComponent implements OnInit {
-  form: FormGroup;
-  roles: string[];
-  userRoleEnum = UserType;
-  classes: Class[];
-  currentRole: string;
-  formControl = new FormControl('', [Validators.required]);
-  EmailFormControl = new FormControl('', [Validators.required, Validators.email]);
-  selectUserType = new FormControl(null, Validators.required);
-  selectGrade = new FormControl('', [Validators.required]);
-  userNameFormControl = new FormControl('', [
-    Validators.required,
-    Validators.minLength(5),
-    Validators.pattern('^[A-Za-z]+$'),
-  ]);
-  constructor(
-    private formBuilder: FormBuilder,
-    private classService: ClassService,
-    public dialogRef: MatDialogRef<AddUserDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: User,
-    public userService: UserService,
-  ) {
-    this.roles = Object.keys(this.userRoleEnum);
-    this.getClasses();
-  }
-
-  ngOnInit(): void {
-    this.form = this.formBuilder.group({
-      firstName: '',
-      lastName: '',
-      userName: this.userNameFormControl,
-      email: this.EmailFormControl,
-      userType: this.selectUserType,
-      class: undefined,
-    });
-  }
-
-  close(): void {
-    this.dialogRef.close();
-  }
-
-  confirmAdd(dialogData): void {
-    this.dialogRef.close(dialogData);
-  }
-
-  onUserTypeChange(event): void {
-    this.currentRole = event;
-    if (event === 'MANAGER') {
-      this.data.Class = undefined;
-    }
-  }
-  async getClasses() {
-    this.classes = await this.classService.getAllClasses();
+export class AddUserDialogComponent extends BaseUserFormDialogComponent implements OnInit {
+  ngOnInit() {
+    this.formOptions.submitButtonLabel = 'הוסף';
   }
 }
