+++ src/app/pages/user/dialogs/update/update-user.dialog.html
@@ -1,70 +1,5 @@
 <div class="container">
-    <h3 mat-dialog-title>עדכון פרטים אישיים</h3>
-    <mat-divider></mat-divider>
-    <form #formControl="ngForm" (submit)="updateDetails()">
-    <mat-dialog-content>
-        <div class="msw-div-column-right">
-          <div class="msw-form-row">
-            <mat-form-field color="primary">
-              <input matInput #input placeholder="שם פרטי" [(ngModel)]="data.firstname" name="id" required >
-              <button type="button" tabindex=-1 mat-button *ngIf="data.firstname" matSuffix mat-icon-button aria-label="Clear" (click)="data.firstname=''">
-                <mat-icon>close</mat-icon>
-              </button>
-                <mat-error *ngIf="formControl.invalid">{{getErrorMessage()}}</mat-error>
-            </mat-form-field>
-        </div>
-        <div class="msw-form-row">
-            <mat-form-field color="primary">
-              <input matInput #input placeholder="שם משתמש" [(ngModel)]="data.username"
-                name="userName" required>
-                <button type="button" tabindex=-1 mat-button *ngIf="data.username" matSuffix mat-icon-button aria-label="Clear" (click)="data.username=''">
-                  <mat-icon>close</mat-icon>
-                </button>
-                <mat-error *ngIf="formControl.invalid">{{getErrorMessage()}}</mat-error>
-            </mat-form-field>
-          </div>
-          <!--TODO:add actual code, this is a dummy code for SCSS updates-->
-          <div class="msw-form-row">
-              <mat-form-field color="primary">
-                <mat-select placeholder="סוג משתמש">
-                  <mat-option>מנהל</mat-option>
-                </mat-select>
-              </mat-form-field>
-            </div>
-          </div>
-          <div class="msw-div-column-left">
-          <div class="msw-form-row">
-            <mat-form-field color="primary">
-              <input matInput #input  placeholder="שם משפחה" [(ngModel)]="data.lastname" name="lastName" required>
-               <button type="button" tabindex=-1 mat-button *ngIf="data.lastname" matSuffix mat-icon-button aria-label="Clear" (click)="data.lastname=''">
-                <mat-icon>close</mat-icon>
-              </button>
-              <mat-error *ngIf="formControl.invalid">{{getErrorMessage()}}</mat-error>
-            </mat-form-field>
-          </div>
-          <div class="msw-form-row">
-            <mat-form-field color="primary">
-              <input matInput #input  placeholder="אימייל" [(ngModel)]="data.email" name="email" required>
-               <button type="button" tabindex=-1 mat-button *ngIf="data.email" matSuffix mat-icon-button aria-label="Clear" (click)="data.email=''">
-                <mat-icon>close</mat-icon>
-              </button>
-              <mat-error *ngIf="formControl.invalid">{{getErrorMessage()}}</mat-error>
-            </mat-form-field>
-          </div>
-          <!--TODO: fix required-->
-          <!--TODO:add actual code, this is a dummy code for SCSS updates-->
-          <div class="msw-form-row">
-            <mat-form-field color="primary">
-              <mat-select placeholder="כיתה" >
-                <mat-option [value]="">בחר</mat-option>
-              </mat-select>
-            </mat-form-field>
-          </div>
-        </div>
-      </mat-dialog-content>
-      <div mat-dialog-actions>
-        <button  mat-button class="msw-button-cancel" tabindex="-1" (click)="close()">ביטול</button>
-        <button type="submit" mat-button class="msw-button-update" [disabled]="!formControl.valid" >עדכן</button>
-      </div>
-    </form>
-  </div>
+  <h3 mat-dialog-title>עדכון פרטים אישיים</h3>
+  <mat-divider></mat-divider>
+  <app-user-details-form [options]="formOptions" [data]="dialogData" (formSubmit)="submitForm($event)" (cancel)="close()"></app-user-details-form>
+</div>
