+++ src/app/pages/user/dialogs/update/update-user.dialog.ts
@@ -1,37 +1,19 @@
-import { Component, Inject } from '@angular/core';
+import { BaseUserFormDialogComponent } from './../base/base-user-form-dialog';
+import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
-import { FormControl, Validators } from '@angular/forms';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { User, UserType } from '../../../../models/user.model';
+import { UserService } from '../../services/user.service';
+import { ClassService } from '../../../class/services/class.graphql.service';
+import { Class } from '../../../../models/class.model';
 
 @Component({
   selector: 'app-update-user.dialog',
   templateUrl: './update-user.dialog.html',
   styleUrls: ['./update-user.dialog.scss'],
 })
-export class UpdateUserDialogComponent {
-  constructor(public dialogRef: MatDialogRef<UpdateUserDialogComponent>, @Inject(MAT_DIALOG_DATA) public data: any) {}
-
-  formControl = new FormControl('', [Validators.required]);
-
-  getErrorMessage() {
-    return this.formControl.hasError('required')
-      ? 'Required field'
-      : this.formControl.hasError('email')
-        ? 'Not a valid email'
-        : '';
-  }
-
-  close(): void {
-    this.dialogRef.close();
-  }
-
-  updateDetails(): void {
-    this.dialogRef.close(this.data);
-  }
-
-  onUserTypeChange(event): void {
-    console.log('class value is: ' + this.data._class);
-    if (event.value === 'MANAGER') {
-      this.data._class = undefined;
-    }
+export class UpdateUserDialogComponent extends BaseUserFormDialogComponent implements OnInit {
+  ngOnInit() {
+    this.formOptions.submitButtonLabel = 'עדכן';
   }
 }
